requirements.txt
# Core Dependencies
openai>=1.5.0
google-generativeai>=0.3.0
requests>=2.28.0
tqdm>=4.64.0
pydantic>=2.0.0

# Parsing and Processing
ast-comments>=1.0.1
ast-decompiler>=0.7.0
jinja2>=3.0.0
markdown>=3.4.0
pygments>=2.14.0

# File Handling
pathlib>=1.0.1
python-magic>=0.4.27
chardet>=5.1.0

# Utilities
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.3.0
click>=8.1.3
colorama>=0.4.6
loguru>=0.7.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-mock>=3.10.0

# Documentation
mkdocs>=1.4.3
mkdocs-material>=9.1.15 